keyvault/python/requirements.txt
azure-identity>=1.14.0
azure-keyvault-secrets>=4.8.0,<=4.10.0
